Discord Bot
Discord bot (nicknamed "The Shooter") written in C# using the Discord.NET Unofficial wrapper.  It features basic moderation features, such as channel renaming and kicking/banning.  The name of the bot was based off of the name of the text channel I was in at the time.  No real thought was poured into it.

The bot was designed as an introduction into C# and Discord bot programming, in general.  I rarely, if ever, found useful tutorials online and wanted to experiment.  At the time, I was trying to build a bot to perform basic commands and then expanded upon it later on.  I'd like to clean up the code and continue working on it at a later date.

I don't mind if you use it in your channel, but it's buggy and I haven't finished refining it just yet.  I'd like for this to be an open tutorial bot for anyone interested in programming Discord bots using C# or the unofficial wrapper.

Planning on:
Adding a mathematics module with LaTeX built in.
This module should be able to solve basic mathematics problems.
Streamlining the moderation features.
Streamlining the help/command menu.
Adding a gaming component to it. ***
(Poker?)
Adding a music portion to it. ***

Will add commands/help documentation soon.

*** NOTE***
The Shooter was supposed to primarily be a greeting bot, hence the name of the file Greetings.cs in src/Modules.  Don't let this fool you, however!  Greetings.cs contains moderation features which will assist you in managing your server.  In fact, a future update will feature a separate module altogether.

- enraged from Aucoustic Mafia